# swagger.yaml
openapi: 3.0.0
info:
  title: Thryve API
  description: |
    **Thryve** is a comprehensive chronic illness community platform API that provides secure authentication, user management, admin controls, and community features.
    
    ## Features
    - 🔐 **JWT-based Authentication** with refresh tokens
    - 👥 **Role-based Access Control** (USER, ADMIN, DOCTOR, MENTOR)
    - 📧 **Email Verification** and password reset
    - 👤 **User Profile Management** with avatar uploads
    - 🔍 **User Search** and public profiles
    - 🏥 **Illness Management** and user preferences
    - 🛡️ **Admin Dashboard** with comprehensive user controls
    - 📊 **Audit Logging** for all admin actions
    
    ## Authentication
    The API uses JWT tokens for authentication:
    - **Access Token**: Short-lived (15 minutes), sent in Authorization header
    - **Refresh Token**: Long-lived (7 days), stored as httpOnly cookie
    
    ## Error Handling
    All endpoints return standardized error responses with appropriate HTTP status codes:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found
    - `409` - Conflict (duplicate data)
    - `422` - Unprocessable Entity (validation errors with details)
    - `500` - Internal Server Error
    
  contact:
    name: Kanishk Chandna
    email: kanishkchandna29@gmail.com
    url: https://github.com/Kanishk2004
  license:
    name: MIT
...
servers:
...

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
  schemas:
...
    ApiResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        data:
          type: object
          description: Response data
        message:
          type: string
          description: Response message
          example: "Operation completed successfully"
        success:
          type: boolean
          description: Whether the operation was successful
          example: true
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        data:
          type: object
          nullable: true
          description: Error data (usually null)
        message:
          type: string
          description: Error message
          example: "Validation failed"
        success:
          type: boolean
          description: Whether the operation was successful
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Invalid email format"

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          description: User's username
          example: "john_doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        fullName:
          type: string
          nullable: true
          description: User's full name
          example: "John Doe"
        bio:
          type: string
          nullable: true
          description: User's biography
          example: "Living with diabetes for 5 years"
        role:
          type: string
          enum: [USER, ADMIN, DOCTOR, MENTOR]
          description: User's role
          example: "USER"
        isEmailVerified:
          type: boolean
          description: Whether email is verified
          example: true
        isActive:
          type: boolean
          description: Whether account is active
          example: true
        isAnonymous:
          type: boolean
          description: Privacy mode status
          example: false
        avatarURL:
          type: string
          nullable: true
          description: URL to user's avatar
          example: "https://res.cloudinary.com/thryve/image/upload/v1234567890/avatar.jpg"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER, PREFER_NOT_TO_SAY]
          nullable: true
          description: User's gender
        dateOfBirth:
          type: string
          format: date
          nullable: true
          description: User's date of birth
          example: "1990-01-15"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-15T10:30:00Z"
...
    AdminAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          example: "USER_BANNED"
        target_id:
          type: string
          format: uuid
        notes:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # Pagination Schema
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        totalUsers:
          type: integer
          example: 95
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
paths:
...
  /admin/users:
    get:
      summary: Get all users (Admin only)
      description: Get paginated list of all users with filtering options
      operationId: getAllUsers
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of users per page
        - in: query
          name: role
          schema:
            type: string
            enum: [ADMIN, DOCTOR, MENTOR, USER]
          description: Filter by user role
        - in: query
          name: verified
          schema:
            type: boolean
          description: Filter by email verification status
        - in: query
          name: isActive
          schema:
            type: boolean
          description: Filter by account active status
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{id}:
    get:
      summary: Get user by ID (Admin only)
      description: Get detailed information about a specific user
      operationId: getUserById
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to fetch
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update user by Admin
      description: Update user information (Admin only)
      operationId: updateUserByAdmin
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  example: "updated_username"
                bio:
                  type: string
                  example: "Updated bio"
                fullName:
                  type: string
                  example: "John Updated"
                gender:
                  type: string
                  enum: [MALE, FEMALE, OTHER, PREFER_NOT_TO_SAY]
                dateOfBirth:
                  type: string
                  format: date
                  example: "1990-01-15"
                isEmailVerified:
                  type: boolean
                  example: true
                isAnonymous:
                  type: boolean
                  example: false
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Validation errors or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete user (Admin only)
      description: Permanently delete a user account
      operationId: deleteUserByAdmin
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to delete
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deletedUser:
                            type: object
                            properties:
                              id:
                                type: string
                                format: uuid
                              username:
                                type: string
                              email:
                                type: string
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required or cannot delete admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{id}/change-role:
    patch:
      summary: Change user role (Admin only)
      description: Change a user's role in the system
      operationId: changeRole
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID whose role to change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [ADMIN, DOCTOR, MENTOR, USER]
                  example: "DOCTOR"
      responses:
        '200':
          description: User role updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Invalid role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{id}/toggle-ban:
    patch:
      summary: Toggle user ban status (Admin only)
      description: Ban or unban a user account
      operationId: toggleUserBan
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to ban/unban
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for ban/unban action
                  example: "Violation of community guidelines"
      responses:
        '200':
          description: User ban status toggled successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/actions/recent:
    get:
      summary: Get recent admin actions
      description: Get list of recent actions performed by the admin
      operationId: getRecentAdminActions
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Recent admin actions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminAction'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'